backend/requirements.txt
flask
flask-cors
joblib
pandas
mysql-connector-python
gunicorn
scikit-learn